backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
scipy==1.12.0
pdf2image==1.17.0
redis==5.0.1
aiofiles==23.2.1
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
wfdb==4.1.2
